+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
+### Fixed
+- year slider to ignore its values if it was not touched
 
 ## [2.63.0] - 2022-08-19
 ### Added