+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file[^1].
 ## [2.63.0] - 2022-08-19
 ### Added
 - item sortables
+- filter items by collection in admin
+- confirm collection delete
 
 ### Changed
 - MySQL database to enabled strict mode