+++ CHANGELOG.md
@@ -3,8 +3,15 @@ All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
 ### Added
-- item sortables
+- importer role
+- import permissions
 - clear image functionality for authorities in admin
+
+## [2.63.0] - 2022-08-19
+### Added
+- item sortables
+- filter items by collection in admin
+- confirm collection delete
 
 ### Changed
 - MySQL database to enabled strict mode